+++ src/integrations/crm/MicrosoftDynamics365.php
@@ -29,6 +29,7 @@ class MicrosoftDynamics365 extends Crm
 
     public const EVENT_MODIFY_REQUIRED_LEVELS = 'modifyRequiredLevels';
     public const EVENT_MODIFY_TARGET_SCHEMAS = 'modifyTargetSchemas';
+    
 
 
     // Static Methods
@@ -67,6 +68,7 @@ class MicrosoftDynamics365 extends Crm
     public ?array $incidentFieldMapping = null;
 
     private array $_entityOptions = [];
+    private array $_systemUsers = [];
 
 
     // Public Methods
@@ -602,6 +604,16 @@ class MicrosoftDynamics365 extends Crm
         // This can be for multiple entities, so have some cache.
         $this->_getEntityOwnerOptions($entity, $fields);
 
+        // Add a list of system users for "Created By"
+        $fields['createdby'] = new IntegrationField([
+            'handle' => 'createdby',
+            'name' => Craft::t('formie', 'Created By'),
+            'options' => [
+                'label' => Craft::t('formie', 'Created By'),
+                'options' => $this->_getSystemUsersOptions(),
+            ],
+        ]);
+
         // Reset array keys
         $fields = array_values($fields);
 
@@ -712,10 +724,11 @@ class MicrosoftDynamics365 extends Crm
                 // Fetch the entities and use the schema options to store. Be sure to limit and be performant.
                 $response = $this->request('GET', $targetSchema['entity'], [
                     'query' => [
-                        '$top' => $targetSchema['limit'] ?? '100',
-                        '$select' => implode(',', $select),
+                        '$expand' => $targetSchema['expand'] ?? null,
+                        '$filter' => $targetSchema['filter'] ?? null,
                         '$orderby' => $targetSchema['orderby'] ?? null,
-                        '$filter' => $targetSchema['filter'] ?? null
+                        '$select' => implode(',', $select),
+                        '$top' => $targetSchema['limit'] ?? '100'
                     ],
                 ]);
 
@@ -801,5 +814,27 @@ class MicrosoftDynamics365 extends Crm
         }
 
         return $path;
+    }
+
+    private function _getSystemUsersOptions(): array
+    {
+        if ($this->_systemUsers) {
+            return $this->_systemUsers;
+        }
+
+        $response = $this->request('GET', 'systemusers', [
+            'query' => [
+                '$top' => '100',
+                '$select' => 'fullname,systemuserid,applicationid',
+                '$orderby' => 'fullname',
+                '$filter' => 'applicationid eq null and invitestatuscode eq 4 and isdisabled eq false',
+            ]
+        ]);
+
+        foreach (($response['value'] ?? []) as $user) {
+            $this->_systemUsers[] = ['label' => $user['fullname'], 'value' => 'systemusers(' . $user['systemuserid'] . ')'];
+        }
+
+        return $this->_systemUsers;
     }
 }