+++ src/services/Fields.php
@@ -379,6 +379,11 @@ class Fields extends Component
                 continue;
             }
 
+            // Is this a non-Formie field, or a MissingField?
+            if (!($field instanceof FormFieldInterface)) {
+                continue;
+            }
+
             // Get the UI
             $uid = $matches['uid'];
 
@@ -470,6 +475,32 @@ class Fields extends Component
         }
 
         return $allFields;
+    }
+
+    public function checkRequiredPlugin(FieldInterface $field): bool
+    {
+        if (!method_exists($field, 'getRequiredPlugins')) {
+            throw new MissingComponentException();
+        }
+
+        foreach ($field::getRequiredPlugins() as $requiredPlugin) {
+            $version = $requiredPlugin['version'] ?? 0;
+            $handle = $requiredPlugin['handle'] ?? '';
+
+            if ($handle) {
+                if (!Formie::$plugin->getService()->isPluginInstalledAndEnabled($handle)) {
+                    throw new MissingComponentException();
+                }
+
+                $plugin = Craft::$app->getPlugins()->getPlugin($handle);
+
+                if (version_compare($plugin->getVersion(), $version, '<')) {
+                    throw new MissingComponentException();
+                }
+            }
+        }
+
+        return true;
     }
 
     /**
@@ -1065,6 +1096,30 @@ class Fields extends Component
             $reservedWords,
             HandleValidator::$baseReservedWords
         );
+    }
+
+    public function getFieldsForForm(Form $form): array
+    {
+        $fields = [];
+
+        if ($fieldLayout =$form->getFormFieldLayout()) {
+            foreach ($fieldLayout->getCustomFields() as $field) {
+                $fields[] = $field;
+
+                // Include any nested fields at the top-level
+                if ($field instanceof NestedFieldInterface) {
+                    if ($nestedFieldLayout = $field->getFieldLayout()) {
+                        foreach ($nestedFieldLayout->getCustomFields() as $nestedField) {
+                            $nestedField->setParentField($field);
+
+                            $fields[] = $nestedField;
+                        }
+                    }
+                }
+            }
+        }
+
+        return $fields;
     }