+++ src/services/Rendering.php
@@ -72,6 +72,9 @@ class Rendering extends Component
             return null;
         }
 
+        // Give the form a unique ID for each render, to help with multiple renders of the same form
+        $form->setFormId($form->getFormId(false));
+
         // Fire a 'modifyFormRenderOptions' event
         $event = new ModifyFormRenderOptionsEvent([
             'form' => $form,
@@ -363,6 +366,16 @@ class Rendering extends Component
             'Invalid country code',
             'Too short',
             'Too long',
+
+            // PayPal
+            'Missing Authorization ID for approval.',
+            'Payment authorized. Finalise the form to complete payment.',
+            'Unable to authorize payment. Please try again.',
+
+            // Opayo
+            'The request timed out.',
+            'The request encountered a network error. Please try again.',
+
         ]);
     }